telemetry: update tests to not use EnvoyFilter
Split out test changes from https://github.com/istio/istio/pull/47892
