+++ pilot/pkg/model/telemetry.go
@@ -1119,6 +1119,7 @@ func generateSDConfig(class networking.ListenerClass, telemetryConfig telemetryF
 	}
 
 	cfg.MetricExpiryDuration = durationpb.New(1 * time.Hour)
+	cfg.EnableAuditLog = features.StackdriverAuditLog
 	// In WASM we are not actually processing protobuf at all, so we need to encode this to JSON
 	cfgJSON, _ := protomarshal.MarshalProtoNames(&cfg)
