+++ pkg/test/fakes/gce_metadata_server/Makefile
@@ -20,7 +20,7 @@ IMG ?= gcr.io/istio-testing/fake-gce-metadata
 
 # NOTE: TAG should be updated whenever changes are made in this directory
 # This should also be updated in dependent components
-TAG := 1.1
+TAG := 1.2
 
 all: build_and_push clean
