+++ tests/integration/telemetry/api/registry_setup_test.go
@@ -23,7 +23,6 @@ import (
 
 	"istio.io/istio/pkg/test/framework/components/registryredirector"
 	"istio.io/istio/pkg/test/framework/resource"
-	"istio.io/istio/tests/integration/telemetry/common"
 )
 
 var registry registryredirector.Instance
@@ -40,15 +39,6 @@ func testRegistrySetup(ctx resource.Context) (err error) {
 	})
 	if err != nil {
 		return
-	}
-
-	args := map[string]any{
-		"DockerConfigJson": base64.StdEncoding.EncodeToString(
-			[]byte(createDockerCredential(registryUser, registryPasswd, registry.Address()))),
-	}
-	if err := ctx.ConfigIstio().EvalFile(common.GetAppNamespace().Name(), args, "testdata/registry-secret.yaml").
-		Apply(); err != nil {
-		return err
 	}
 
 	return nil
