--- tests/integration/telemetry/envoyfilter/customizemetrics/testdata/setup_config.yaml
@@ -1,35 +1,0 @@
-meshConfig:
-  defaultConfig:
-    proxyStatsMatcher:
-      inclusionPrefixes:
-      - istio_custom_total
-    extraStatTags:
-    - url_path
-    - response_status
-values:
- telemetry:
-   v2:
-     prometheus:
-       configOverride:
-         inboundSidecar:
-           debug: false
-           stat_prefix: istio
-           metrics:
-           - name: requests_total
-             dimensions:
-               response_code: istio_responseClass
-               request_operation: istio_operationId
-               grpc_response_status: istio_grpcResponseStatus
-               custom_dimension: "'test'"
-             tags_to_remove:
-             - "%s"
-         outboundSidecar:
-           definitions:
-           - name: custom_total
-             type: "COUNTER"
-             value: "1"
-           metrics:
-           - name: custom_total
-             dimensions:
-               url_path: request.url_path
-               response_status: string(response.code)
