--- tests/integration/telemetry/envoyfilter/nullvm/stats_tcp_filter_test.go
@@ -1,32 +1,0 @@
-//go:build integ
-// +build integ
-
-// Copyright Istio Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package nullvm
-
-import (
-	"testing"
-
-	"istio.io/istio/tests/integration/telemetry/common"
-)
-
-func TestTcpMetric(t *testing.T) {
-	common.TestStatsTCPFilter(t, "observability.telemetry.stats.prometheus.tcp")
-}
-
-func TestGatewayTCP(t *testing.T) {
-	common.TestStatsGatewayServerTCPFilter(t, "observability.telemetry.stats.prometheus.tcp")
-}
