+++ tests/integration/telemetry/stackdriver/testdata/security_authz_dry_run/server_access_log_deny_no_policy.json.tmpl
@@ -19,7 +19,7 @@
       "source_namespace": "{{ .EchoNamespace }}",
       "destination_service_name": "srv",
       "destination_version": "v1",
-      "log_sampled": "true",
+      "log_sampled": "false",
       "mesh_uid": "proj-test-mesh",
       "protocol": "http",
       "requested_server_name": "outbound_.80_._.srv.{{ .EchoNamespace }}.svc.cluster.local",
