+++ tests/integration/telemetry/stackdriver/testdata/server_access_log.json.tmpl
@@ -26,7 +26,7 @@
         "source_version": "v1",
         "source_workload": "clt-{{ .ClusterName }}-v1",
         "protocol": "grpc",
-        "log_sampled": "true",
+        "log_sampled": "false",
         "upstream_cluster": "inbound|7070||",
         "route_name": "default",
         "requested_server_name": "outbound_.7070_._.srv.{{ .EchoNamespace }}.svc.cluster.local",
